장비해제, 장비 능력치 적용, 골드증가 적용

+++ Assets/Scripts/DataTables/UpgradeTable.cs
@@ -13,7 +13,7 @@ public class UpgradeData
     public int DivValue { get; set; } 
     public int Gold { get; set; }
     public int GoldRange { get; set; }
-    public float GoldincreaseValue { get; set; }
+    public int GoldincreaseValue { get; set; }
     public int MaxLv { get; set; }
     public string UpgradeIcon { get; set; }
     public int StringID { get; set; }
